requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==2.3.3
flatbuffers==24.3.7
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
keras==2.13.1
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==2.1.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
# tensorflow-macos==2.13.0
termcolor==2.4.0
typing_extensions==4.5.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
Pillow==10.0.0
